feat: add toggle
- update the app header
- create listItemDetails
- create swag gird item
