+++ src/js/components/SwagGridItem.js
@@ -1,88 +1,33 @@
 import React, { Component } from 'react';
-import { Image, StyleSheet, Text, TouchableOpacity, View } from 'react-native';
-import { MAKE_ACCESSIBLE_FOR_AUTOMATION, MUSEO_SANS_BOLD, MUSEO_SANS_NORMAL, WINDOW_WIDTH } from '../config/Constants';
 import { ShoppingCart } from '../shopping-cart';
 import { Button, Divider } from 'react-native-elements';
-import { colors } from '../utils/colors';
-import { testProperties } from '../config/TestProperties';
 import i18n from '../config/i18n';
-import { Credentials } from '../credentials';
+import { testProperties } from '../config/TestProperties';
+import { Image, StyleSheet, Text, TouchableOpacity, View } from 'react-native';
+import {
+  MAKE_ACCESSIBLE_FOR_AUTOMATION,
+  MUSEO_SANS_BOLD,
+  MUSEO_SANS_NORMAL,
+  WINDOW_WIDTH,
+} from '../config/Constants';
+import { colors } from '../utils/colors';
 
-export default class InventoryListItemColumn extends Component {
+export default class SwagGridItem extends Component {
   constructor(props) {
     super(props);
-
-    this.state = {
-      id: props.id,
-      image_url: props.image_url,
-      name: props.name,
-      desc: props.desc,
-      price: props.price,
-      // Set our initial state now
-      itemInCart: ShoppingCart.isItemInCart(props.id),
-    };
-
-    ShoppingCart.registerCartListener(this);
-
-    if (Credentials.isProblemUser()) {
-      // Replace our image with our broken link image
-      this.state.image_url = require('../../img/sl-404.jpg');
-    }
-
-    // This prop will disable the navigation / flashy press if it should not be touchable
-    this.isDisabled = this.props.disabled || false;
-
-    // Need to pass this in explicitly since it's a subcomponent
-    this.navigation = props.navigation;
-
-    this.addToCart = this.addToCart.bind(this);
-    this.removeFromCart = this.removeFromCart.bind(this);
-    this.navigateToItem = this.navigateToItem.bind(this);
-  }
-
-  addToCart() {
-    if (Credentials.isProblemUser()) {
-      // Bail out now, don't add to cart if the item ID is odd
-      if (this.state.id % 2 === 1) {
-        return;
-      }
-    }
-
-    ShoppingCart.addItem(this.state.id);
-    this.setState({ itemInCart: true });
-  }
-
-  removeFromCart() {
-    if (Credentials.isProblemUser()) {
-      // Bail out now, don't remove from cart if the item ID is even
-      if (this.state.id % 2 === 0) {
-        return;
-      }
-    }
-
-    ShoppingCart.removeItem(this.state.id);
-    this.setState({ itemInCart: false });
-  }
-
-  navigateToItem() {
-    let itemId = this.state.id;
-    if (Credentials.isProblemUser()) {
-      itemId += 1;
-    }
-
-    this.navigation.navigate('InventoryItem', { id: itemId });
   }
 
   render() {
     let cartButton;
+    const { addToCart, id, index, image_url, name, navigateToItem, price, removeFromCart } = this.props;
 
-    if (ShoppingCart.isItemInCart(this.state.id)) {
+    if (ShoppingCart.isItemInCart(id)) {
       cartButton = (
         <Button
           buttonStyle={ [ styles.button_style, styles.remove_button_style ] }
           containerStyle={ styles.button_container_style }
           titleStyle={ [ styles.button_title_style, styles.remove_button_title_style ] }
-          onPress={ this.removeFromCart }
+          onPress={ removeFromCart }
           title={ i18n.t('inventoryItemPage.removeButton') }
           { ...testProperties(i18n.t('inventoryListPage.removeButton')) }
         />);
@@ -92,13 +37,11 @@ export default class InventoryListItemColumn extends Component {
           buttonStyle={ styles.button_style }
           containerStyle={ styles.button_container_style }
           titleStyle={ styles.button_title_style }
-          onPress={ this.addToCart }
+          onPress={ addToCart }
           title={ i18n.t('inventoryItemPage.addButton') }
           { ...testProperties(i18n.t('inventoryListPage.addButton')) }
         />);
     }
-
-    const { index } = this.props;
 
     return (
       <View
@@ -107,14 +50,13 @@ export default class InventoryListItemColumn extends Component {
         { ...testProperties(i18n.t('inventoryListPage.itemContainer')) }
       >
         <TouchableOpacity
-          onPress={ this.navigateToItem }
+          onPress={ navigateToItem }
           { ...MAKE_ACCESSIBLE_FOR_AUTOMATION }
-          disabled={ this.isDisabled }
           style={ styles.item_wrapper }
         >
           <View style={ styles.top_container }>
             <Image
-              source={ this.state.image_url }
+              source={ image_url }
               style={ styles.item_image }
               resizeMode="contain"
             />
@@ -122,14 +64,14 @@ export default class InventoryListItemColumn extends Component {
             <Text
               style={ styles.item_name }
               { ...testProperties(i18n.t('inventoryListPage.itemDescription')) }
-            >{ this.state.name }</Text>
+            >{ name }</Text>
           </View>
 
           <View style={ styles.bottom_container }>
             <Divider style={ styles.divider }/>
 
             <Text style={ styles.price_text }{ ...testProperties(i18n.t('inventoryListPage.price')) }>
-              ${ this.state.price }
+              ${ price }
             </Text>
 
             { cartButton }
@@ -153,13 +95,11 @@ const styles = StyleSheet.create({
   item_wrapper: {
     flex: 1,
   },
-
   item_name: {
     color: colors.slRed,
     fontSize: 18,
     fontFamily: MUSEO_SANS_BOLD,
   },
-
   item_image: {
     flex: 1,
     // This is for keeping the aspect ratio and make it responsive
@@ -167,7 +107,6 @@ const styles = StyleSheet.create({
     width: '100%',
     marginBottom: 20,
   },
-
   top_container: {
     flex: 1,
     justifyContent: 'flex-start',
@@ -176,7 +115,6 @@ const styles = StyleSheet.create({
     flex: 1,
     justifyContent: 'flex-end',
   },
-
   divider: {
     borderBottomColor: colors.lightGray,
     borderBottomWidth: 2,
@@ -184,13 +122,11 @@ const styles = StyleSheet.create({
     marginBottom: 10,
     marginTop: 20,
   },
-
   price_text: {
     color: colors.gray,
     fontSize: 22,
     fontFamily: MUSEO_SANS_NORMAL,
   },
-
   button_style: {
     backgroundColor: colors.white,
     borderColor: colors.slRed,
