+++ src/js/screens/InventoryList.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { StyleSheet, FlatList, View, ScrollView, Image } from 'react-native';
+import { StyleSheet, FlatList, View, ScrollView, Image, TouchableOpacity } from 'react-native';
 import { ThemeProvider } from 'react-native-elements';
 import ModalSelector from 'react-native-modal-selector';
 import { InventoryData } from '../data/inventory-data.js';
@@ -8,7 +8,9 @@ import i18n from '../config/i18n';
 import { testProperties } from '../config/TestProperties';
 import { colors } from '../utils/colors';
 import Footer from '../components/Footer';
-import InventoryListItemColumn from '../components/InventoryListItemColumn';
+import InventoryListItem from '../components/InventoryListItem';
+import toggleRow from '../../img/toggle-row.png';
+import toggleGrid from '../../img/toggle-grid.png';
 
 export default class InventoryList extends Component {
   constructor(props) {
@@ -18,6 +20,7 @@ export default class InventoryList extends Component {
       inventoryList: InventoryData.ITEMS_NAME_AZ,
       sortState: 'az',
       menuOpen: false,
+      gridView: true,
     };
 
     this.changeSort = this.changeSort.bind(this);
@@ -27,6 +30,8 @@ export default class InventoryList extends Component {
     this.sortPriceHiLo = this.sortPriceHiLo.bind(this);
     this.openMenu = this.openMenu.bind(this);
     this.closeMenu = this.closeMenu.bind(this);
+    this.toggleState = this.toggleState.bind(this);
+    this.toggleImage = this.toggleImage.bind(this);
   }
 
   changeSort(sortType) {
@@ -77,15 +82,25 @@ export default class InventoryList extends Component {
   }
 
   openMenu() {
-    this.setState({
-      menuOpen: true,
-    });
+    this.setState({ menuOpen: true });
   }
 
   closeMenu() {
-    this.setState({
-      menuOpen: false,
-    });
+    this.setState({ menuOpen: false });
+  }
+
+  toggleState() {
+    this.setState({ gridView: !this.state.gridView });
+  }
+
+  toggleImage() {
+    return (
+      <Image
+        style={ styles.toggle_image }
+        resizeMode="contain"
+        source={ this.state.gridView ? toggleRow : toggleGrid }
+      />
+    );
   }
 
   keyExtractor = item => item.name;
@@ -98,28 +113,34 @@ export default class InventoryList extends Component {
       { key: 'lohi', label: i18n.t('modalSelector.loHiLabel') },
       { key: 'hilo', label: i18n.t('modalSelector.hiLoLabel') },
     ];
-    const modalSelector = (
-      <View
-        { ...testProperties(i18n.t('modalSelector.button')) }
-        style={ styles.selectorContainer }
-      >
-        <ModalSelector
-          data={ sortOptions }
-          style={ styles.selector }
-          selectTextStyle={ styles.selector_text }
-          onChange={ (sortOption) => this.changeSort(sortOption.key) }
-          cancelText={ i18n.t('modalSelector.cancel') }
-          listItemAccessible
-          cancelButtonAccessible
-          openButtonContainerAccessible
-          scrollViewAccessibilityLabel={ i18n.t('modalSelector.container') }
+    const headerButtons = (
+      <View style={ styles.component_container }>
+        <TouchableOpacity
+          style={ styles.toggle }
+          { ...testProperties(i18n.t('inventoryListPage.toggle')) }
+          onPress={ this.toggleState }
         >
-          <Image
-            style={ styles.selector_image }
-            resizeMode="contain"
-            source={ require('../../img/filter-button.png') }
-          />
-        </ModalSelector>
+          { this.toggleImage() }
+        </TouchableOpacity>
+        <View { ...testProperties(i18n.t('modalSelector.button')) }>
+          <ModalSelector
+            data={ sortOptions }
+            style={ styles.selector }
+            selectTextStyle={ styles.selector_text }
+            onChange={ (sortOption) => this.changeSort(sortOption.key) }
+            cancelText={ i18n.t('modalSelector.cancel') }
+            listItemAccessible
+            cancelButtonAccessible
+            openButtonContainerAccessible
+            scrollViewAccessibilityLabel={ i18n.t('modalSelector.container') }
+          >
+            <Image
+              style={ styles.selector_image }
+              resizeMode="contain"
+              source={ require('../../img/filter-button.png') }
+            />
+          </ModalSelector>
+        </View>
       </View>
     );
 
@@ -128,7 +149,7 @@ export default class InventoryList extends Component {
         <AppHeader
           navigation={ this.props.navigation }
           header={ i18n.t('inventoryListPage.header') }
-          component={ modalSelector }
+          component={ headerButtons }
         >
           <ScrollView
             style={ styles.container }
@@ -138,17 +159,21 @@ export default class InventoryList extends Component {
             <FlatList
               data={ this.state.inventoryList }
               keyExtractor={ this.keyExtractor }
-              renderItem={ ({ item, index }) => <InventoryListItemColumn
-                key={ item.id }
-                id={ item.id }
-                image_url={ item.image_url }
-                name={ item.name }
-                desc={ item.desc }
-                price={ item.price }
-                navigation={ this.props.navigation }
-                index={index}
-              /> }
-              numColumns={ 2 }
+              key={ (this.state.gridView) ? 1 : 0 }
+              numColumns={ this.state.gridView ? 2 : 1 }
+              renderItem={ ({ item, index }) =>
+                <InventoryListItem
+                  key={ item.id }
+                  id={ item.id }
+                  image_url={ item.image_url }
+                  name={ item.name }
+                  desc={ item.desc }
+                  price={ item.price }
+                  navigation={ this.props.navigation }
+                  index={ index }
+                  gridView={ this.state.gridView }
+                />
+              }
             />
             <Footer/>
           </ScrollView>
@@ -159,10 +184,16 @@ export default class InventoryList extends Component {
 }
 
 const styles = StyleSheet.create({
-  selectorContainer: {
-    position: 'absolute',
-    right: 10,
-    top: 14,
+  component_container: {
+    flexDirection: 'row',
+    justifyContent: 'flex-end',
+  },
+  toggle: {
+    marginRight: 10,
+  },
+  toggle_image: {
+    height: 40,
+    width: 40,
   },
   selector: {
     height: 40,
